Using original Materials classes instead of new Keyring classes
Since the CryptoMaterialsManager interface uses the original
EncryptionMaterials and DecryptionMaterials, we would force customers
to update their code once we deprecate those materials in favor of the
new Keyring versions. To avoid this, Keyrings will now use modified
versions of the original Materials classes, with certain methods
deprecated.

In addition, this commit removed the MasterKeyProviderKeyring, which
was found to not be feasible for the Java ESDK since MasterKeys are
provided in the result object, which would force MasterKeys to be
incorporated into Keyrings.